Refactored - pulled functions into heper method
